rename file to `types.ts`
