+++ packages/electron-updater/src/AppImageUpdater.ts
@@ -8,7 +8,7 @@ import { BaseUpdater, InstallOptions } from "./BaseUpdater"
 import { DifferentialDownloaderOptions } from "./differentialDownloader/DifferentialDownloader"
 import { FileWithEmbeddedBlockMapDifferentialDownloader } from "./differentialDownloader/FileWithEmbeddedBlockMapDifferentialDownloader"
 import { findFile, Provider } from "./providers/Provider"
-import { DOWNLOAD_PROGRESS, ResolvedUpdateFileInfo } from "./exports"
+import { DOWNLOAD_PROGRESS, ResolvedUpdateFileInfo } from "./types"
 
 export class AppImageUpdater extends BaseUpdater {
   constructor(options?: AllPublishOptions | null, app?: any) {
