+++ packages/electron-updater/src/DebUpdater.ts
@@ -3,7 +3,7 @@ import { AppAdapter } from "./AppAdapter"
 import { DownloadUpdateOptions } from "./AppUpdater"
 import { BaseUpdater, InstallOptions } from "./BaseUpdater"
 import { findFile } from "./providers/Provider"
-import { DOWNLOAD_PROGRESS } from "./exports"
+import { DOWNLOAD_PROGRESS } from "./types"
 
 export class DebUpdater extends BaseUpdater {
   constructor(options?: AllPublishOptions | null, app?: AppAdapter) {
