+++ packages/electron-updater/src/DownloadedUpdateHelper.ts
@@ -3,8 +3,8 @@ import { createHash } from "crypto"
 import { createReadStream } from "fs"
 // @ts-ignore
 import * as isEqual from "lodash.isequal"
-import { ResolvedUpdateFileInfo } from "./exports"
-import { Logger } from "./exports"
+import { ResolvedUpdateFileInfo } from "./types"
+import { Logger } from "./types"
 import { pathExists, readJson, emptyDir, outputJson, unlink } from "fs-extra"
 import * as path from "path"
