+++ packages/electron-updater/src/MacUpdater.ts
@@ -6,7 +6,7 @@ import { createServer, IncomingMessage, Server, ServerResponse } from "http"
 import { AppAdapter } from "./AppAdapter"
 import { AppUpdater, DownloadUpdateOptions } from "./AppUpdater"
 import { ResolvedUpdateFileInfo } from "./main"
-import { UpdateDownloadedEvent } from "./exports"
+import { UpdateDownloadedEvent } from "./types"
 import { findFile } from "./providers/Provider"
 import AutoUpdater = Electron.AutoUpdater
 import { execFileSync } from "child_process"
