+++ packages/electron-updater/src/differentialDownloader/DifferentialDownloader.ts
@@ -4,7 +4,7 @@ import { close, open } from "fs-extra"
 import { createWriteStream } from "fs"
 import { OutgoingHttpHeaders, RequestOptions } from "http"
 import { ProgressInfo, CancellationToken } from "builder-util-runtime"
-import { Logger } from "../exports"
+import { Logger } from "../types"
 import { copyData } from "./DataSplitter"
 import { URL } from "url"
 import { computeOperations, Operation, OperationKind } from "./downloadPlanBuilder"
