+++ packages/electron-updater/src/types.ts
